@@ -6,7 +6,7 @@ authors = ["Tom Zehle, Moritz Schlager, Timo Heiß"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = ">=3.10,<3.13"
 numpy = "^1.26.0"
 pandas = "^2.2.2"
 tqdm = "^4.66.5"
@@ -18,7 +18,7 @@ openai = "^1.0.0"
 requests = "^2.31.0"
 
 [tool.poetry.group.vllm.dependencies]
-vllm = "^0.8.3"
+vllm = "^0.10.1.1"
 
 [tool.poetry.group.transformers.dependencies]
 transformers = "^4.48.0"
@@ -29,13 +29,14 @@ flake8 = "^7.1.0"
 isort = "^5.13.2"
 pre-commit = "^3.7.1"
 ipykernel = "^6.29.5"
+mypy = "^1.8.0"
 
 [tool.poetry.group.test.dependencies]
 pytest = "^8.3.5"
 pytest-cov = "^6.1.1"
 openai = "^1.0.0"
 requests = "^2.31.0"
-vllm = "^0.8.2"
+vllm = "^0.10.1.1"
 transformers = "^4.48.0"
 
 [tool.poetry.group.docs.dependencies]
@@ -62,3 +63,11 @@ lines_between_sections = 1
 
 [tool.pydocstyle]
 convention = "google"
+
+[tool.mypy]
+ignore_missing_imports = true
+disable_error_code = ["call-overload", "assignment", "overload-overlap"]
+
+[[tool.mypy.overrides]]
+module = "transformers.*"
+ignore_errors = true
