@@ -1,17 +1,17 @@
 """Module to interface with various language models through their respective APIs."""
 
-
 try:
     import asyncio
 
     from openai import AsyncOpenAI
+    from openai.types.chat import ChatCompletion, ChatCompletionMessageParam
 
     import_successful = True
 except ImportError:
     import_successful = False
 
 
-from typing import TYPE_CHECKING, List
+from typing import TYPE_CHECKING, Dict, List, Optional
 
 from promptolution.llms.base_llm import BaseLLM
 
@@ -23,9 +23,21 @@ from promptolution.utils.logging import get_logger
 logger = get_logger(__name__)
 
 
-async def _invoke_model(prompt, system_prompt, max_tokens, model_id, client, semaphore, max_retries=20, retry_delay=5):
+async def _invoke_model(
+    prompt: str,
+    system_prompt: str,
+    max_tokens: int,
+    model_id: str,
+    client: AsyncOpenAI,
+    semaphore: asyncio.Semaphore,
+    max_retries: int = 20,
+    retry_delay: float = 5,
+) -> ChatCompletion:
     async with semaphore:
-        messages = [{"role": "system", "content": system_prompt}, {"role": "user", "content": prompt}]
+        messages: List[ChatCompletionMessageParam] = [
+            {"role": "system", "content": system_prompt},
+            {"role": "user", "content": prompt},
+        ]
 
         for attempt in range(max_retries + 1):  # +1 for the initial attempt
             try:
@@ -46,7 +58,8 @@ async def _invoke_model(prompt, system_prompt, max_tokens, model_id, client, sem
                 else:
                     # Log the final failure and re-raise the exception
                     logger.error(f"❌ API call failed after {max_retries + 1} attempts: {str(e)}")
-                    raise
+                    raise  # Re-raise the exception after all retries fail
+        raise RuntimeError("Failed to get response after multiple retries.")
 
 
 class APILLM(BaseLLM):
@@ -65,13 +78,13 @@ class APILLM(BaseLLM):
 
     def __init__(
         self,
-        api_url: str = None,
-        model_id: str = None,
-        api_key: str = None,
-        max_concurrent_calls=50,
-        max_tokens=512,
-        config: "ExperimentConfig" = None,
-    ):
+        api_url: Optional[str] = None,
+        model_id: Optional[str] = None,
+        api_key: Optional[str] = None,
+        max_concurrent_calls: int = 50,
+        max_tokens: int = 512,
+        config: Optional["ExperimentConfig"] = None,
+    ) -> None:
         """Initialize the APILLM with a specific model and API configuration.
 
         Args:
@@ -103,14 +116,26 @@ class APILLM(BaseLLM):
 
     def _get_response(self, prompts: List[str], system_prompts: List[str]) -> List[str]:
         # Setup for async execution in sync context
-        loop = asyncio.get_event_loop()
+        try:
+            loop = asyncio.get_running_loop()
+        except RuntimeError:  # 'get_running_loop' raises a RuntimeError if there is no running loop
+            loop = asyncio.new_event_loop()
+            asyncio.set_event_loop(loop)
+
         responses = loop.run_until_complete(self._get_response_async(prompts, system_prompts))
         return responses
 
     async def _get_response_async(self, prompts: List[str], system_prompts: List[str]) -> List[str]:
+        assert self.model_id is not None, "model_id must be set"
         tasks = [
             _invoke_model(prompt, system_prompt, self.max_tokens, self.model_id, self.client, self.semaphore)
             for prompt, system_prompt in zip(prompts, system_prompts)
         ]
-        responses = await asyncio.gather(*tasks)
-        return [response.choices[0].message.content for response in responses]
+        messages = await asyncio.gather(*tasks)
+        responses = []
+        for message in messages:
+            response = message.choices[0].message.content
+            if response is None:
+                raise ValueError("Received None response from the API.")
+            responses.append(response)
+        return responses
