@@ -3,11 +3,13 @@
 
 from abc import ABC, abstractmethod
 
-from typing import TYPE_CHECKING, Callable, List, Literal
+from typing import TYPE_CHECKING, List, Literal, Optional
 
 if TYPE_CHECKING:  # pragma: no cover
     from promptolution.tasks.base_task import BaseTask
+    from promptolution.predictors.base_predictor import BasePredictor
     from promptolution.utils.config import ExperimentConfig
+    from promptolution.utils.callbacks import BaseCallback
 
 from promptolution.utils.logging import get_logger
 
@@ -24,33 +26,34 @@ class BaseOptimizer(ABC):
     Attributes:
         config (ExperimentConfig, optional): Configuration for the optimizer, overriding defaults.
         prompts (List[str]): List of current prompts being optimized.
-        task (BaseTask): The task object used for evaluating prompts.
+        task (BaseTask): The task object for evaluating prompts.
         callbacks (List[Callable]): List of callback functions to be called during optimization.
         predictor: The predictor used for prompt evaluation (if applicable).
     """
 
     def __init__(
         self,
-        predictor,
+        predictor: "BasePredictor",
         task: "BaseTask",
-        initial_prompts: List[str],
-        callbacks: List[Callable] = None,
-        config: "ExperimentConfig" = None,
-    ):
+        initial_prompts: Optional[List[str]] = None,
+        callbacks: Optional[List["BaseCallback"]] = None,
+        config: Optional["ExperimentConfig"] = None,
+    ) -> None:
         """Initialize the optimizer with a configuration and/or direct parameters.
 
         Args:
-            initial_prompts: Initial set of prompts to start optimization with.
             task: Task object for prompt evaluation.
-            callbacks: List of callback functions.
             predictor: Predictor for prompt evaluation.
+            initial_prompts: Initial set of prompts to start optimization with.
+            callbacks: List of callback functions.
             config (ExperimentConfig, optional): Configuration for the optimizer, overriding defaults.
         """
         # Set up optimizer state
-        self.prompts = initial_prompts
+        self.prompts: List[str] = initial_prompts or []
         self.task = task
-        self.callbacks = callbacks or []
+        self.callbacks: List["BaseCallback"] = callbacks or []
         self.predictor = predictor
+        self.scores: List[float] = []
 
         if config is not None:
             config.apply_to(self)
@@ -81,7 +84,7 @@ class BaseOptimizer(ABC):
                 # exit training loop and gracefully fail
                 logger.error(f"⛔ Error during optimization step: {e}")
                 logger.error("⚠️ Exiting optimization loop.")
-                continue_optimization = False
+                break
 
             # Callbacks at the end of each step
             continue_optimization = self._on_step_end()
@@ -93,7 +96,7 @@ class BaseOptimizer(ABC):
         return self.prompts
 
     @abstractmethod
-    def _pre_optimization_loop(self):
+    def _pre_optimization_loop(self) -> None:
         """Prepare for the optimization loop.
 
         This method should be implemented by concrete optimizer classes to define
@@ -113,15 +116,16 @@ class BaseOptimizer(ABC):
         """
         pass
 
-    def _on_step_end(self):
+    def _on_step_end(self) -> bool:
         """Call all registered callbacks at the end of each optimization step."""
         continue_optimization = True
         for callback in self.callbacks:
-            continue_optimization &= callback.on_step_end(self)  # if any callback returns False, end the optimization
+            if not callback.on_step_end(self):
+                continue_optimization = False
 
         return continue_optimization
 
-    def _on_train_end(self):
+    def _on_train_end(self) -> None:
         """Call all registered callbacks at the end of the entire optimization process."""
         for callback in self.callbacks:
             callback.on_train_end(self)
