@@ -1,8 +1,18 @@
 """Utils for formatting prompts and outputs."""
-from typing import List, Union
+from typing import List, Union, overload
 
 
-def extract_from_tag(text: str, start_tag: str, end_tag: str) -> Union[List[str], str]:
+@overload
+def extract_from_tag(text: str, start_tag: str, end_tag: str) -> str:
+    ...
+
+
+@overload
+def extract_from_tag(text: List[str], start_tag: str, end_tag: str) -> List[str]:
+    ...
+
+
+def extract_from_tag(text: Union[str, List[str]], start_tag: str, end_tag: str) -> Union[List[str], str]:
     """Extracts content from a string between specified start and end tags.
 
     Args:
