@@ -36,18 +36,17 @@ class MockPredictor(BasePredictor):
         self.predetermined_predictions = predetermined_predictions or {}
         self.call_history = []
 
-    def _extract_preds(self, preds: List[str], shape: Tuple[int, int] = None) -> np.ndarray:
+    def _extract_preds(self, preds: List[str]) -> List[str]:
         """Extract predictions based on predetermined mapping or default behavior.
 
         Args:
             preds: Raw text predictions
-            shape: Shape for reshaping results (optional)
 
         Returns:
-            np.ndarray: Extracted predictions
+           List[str]: Extracted predictions
         """
         # Record call for test assertions
-        self.call_history.append({"preds": preds, "shape": shape})
+        self.call_history.append({"preds": preds})
 
         results = []
         for pred in preds:
@@ -58,9 +57,5 @@ class MockPredictor(BasePredictor):
                 results.append(self.classes[0])
 
         results_array = np.array(results)
-
-        # Reshape if shape is provided
-        if shape is not None:
-            results_array = results_array.reshape(shape)
 
         return results_array
