@@ -47,89 +47,23 @@ class MockTask(BaseTask):
         self.increment_block_idx = MagicMock()
         self.reset_block_idx = MagicMock()
 
-    def _evaluate(self, x: np.ndarray, y: np.ndarray, pred: np.ndarray, **kwargs) -> float:
+    def _evaluate(self, xs: List[str], ys: List[str], preds: List[str], **kwargs) -> List[float]:
         """Calculate the score for a single prediction.
 
         Args:
-            x: Input data (not used in mock)
-            y: Ground truth labels (not used in mock)
-            pred: Predicted labels
+            xs: Input data (not used in mock)
+            ys: Ground truth labels (not used in mock)
+            preds: Predicted labels
 
         Returns:
             Score based on predetermined scores or a default logic.
         """
         if isinstance(self.predetermined_scores, dict):
-            return self.predetermined_scores.get(pred, 0.5)
+            return [self.predetermined_scores.get(pred, 0) for pred in preds]
         elif isinstance(self.predetermined_scores, list):
             self.score_index += 1
-            return self.predetermined_scores[(self.score_index - 1) % len(self.predetermined_scores)]
+            return self.predetermined_scores
         elif callable(self.predetermined_scores):
-            return self.predetermined_scores(x)
+            return self.predetermined_scores(xs)
         else:
-            return len(pred)
-
-    # def evaluate(
-    #     self,
-    #     prompts: List[str],
-    #     predictor,
-    #     eval_strategy: str = "subsample",
-    #     system_prompts: List[str] = None,
-    #     return_agg_scores: bool = False,
-    #     return_seq: bool = False,
-    # ) -> np.ndarray:
-    #     """Evaluate prompts with predetermined scores.
-
-    #     Args:
-    #         prompts: List of prompts to evaluate
-    #         predictor: Predictor (ignored in mock)
-    #         system_prompts: System prompts (ignored in mock)
-    #         subsample: Whether to subsample (ignored in mock)
-    #         n_samples: Number of samples (ignored in mock)
-    #         return_seq: Whether to return sequences
-
-    #     Returns:
-    #         np.ndarray of scores, and optionally sequences
-    #     """
-    #     # Record the call
-    #     self.call_history.append(
-    #         {
-    #             "prompts": prompts,
-    #             "predictor": predictor,
-    #             "system_prompts": system_prompts,
-    #             "eval_strategy": eval_strategy,
-    #             "return_agg_scores": return_agg_scores,
-    #             "return_seq": return_seq,
-    #         }
-    #     )
-
-    #     scores = []
-    #     for prompt in prompts:
-    #         # Handle different types of predetermined_scores
-    #         if callable(self.predetermined_scores):
-    #             # If it's a function, call it with the prompt
-    #             score = self.predetermined_scores(prompt)
-    #         elif isinstance(self.predetermined_scores, dict) and prompt in self.predetermined_scores:
-    #             # If it's a dict, look up the prompt
-    #             score = self.predetermined_scores[prompt]
-    #         elif isinstance(self.predetermined_scores, list):
-    #             # If it's a list, return items in sequence (cycling if needed)
-    #             if self.score_index < len(self.predetermined_scores):
-    #                 score = self.predetermined_scores[self.score_index]
-    #                 self.score_index = (self.score_index + 1) % len(self.predetermined_scores)
-    #             else:
-    #                 score = 0.5  # Default score
-    #         else:
-    #             # Generate a somewhat predictable score based on prompt length
-    #             # (longer prompts get slightly higher scores)
-    #             score = 0.5 + 0.01 * (len(prompt) % 10)
-
-    #         scores.append(score)
-
-    #     scores_array = np.array(scores)
-
-    #     if return_seq:
-    #         # Generate dummy sequences
-    #         seqs = [[f"Input: {x}\nOutput: {prompt}" for x in self.xs] for prompt in prompts]
-    #         return scores_array, seqs
-
-    #     return scores_array
+            return [len(pred) for pred in preds]
