@@ -8,7 +8,7 @@ def test_reward_task_initialization(mock_reward_task, simple_reward_function):
     assert mock_reward_task.x_column == "x"
     assert not mock_reward_task.has_y
     assert len(mock_reward_task.xs) == len(mock_reward_task.df)
-    assert np.all(mock_reward_task.ys == None)  # noqa: E711
+    assert all(y == "" for y in mock_reward_task.ys)  # noqa: E711
 
 
 def test_reward_task_initialization_no_x_column(mock_reward_task_no_x_column, simple_reward_function):
@@ -17,7 +17,7 @@ def test_reward_task_initialization_no_x_column(mock_reward_task_no_x_column, si
     assert not mock_reward_task_no_x_column.has_y
     assert len(mock_reward_task_no_x_column.xs) == len(mock_reward_task_no_x_column.df)
     assert all(x == "" for x in mock_reward_task_no_x_column.xs)
-    assert np.all(mock_reward_task_no_x_column.ys == None)  # noqa: E711
+    assert all([y == "" for y in mock_reward_task_no_x_column.ys])  # noqa: E711
 
 
 def test_reward_task_evaluate_with_return_seq(mock_reward_task, mock_predictor):
@@ -26,9 +26,5 @@ def test_reward_task_evaluate_with_return_seq(mock_reward_task, mock_predictor):
 
     scores, seqs = mock_reward_task.evaluate(prompts, mock_predictor, return_seq=True, return_agg_scores=False)
 
-    assert scores.shape == (1, len(mock_reward_task.xs))
+    assert len(scores) == 1
     assert len(seqs) == 1
-    assert len(seqs[0]) == len(mock_reward_task.xs)
-
-    for seq in seqs[0]:
-        assert any(str(x_val) in seq for x_val in mock_reward_task.xs)
